requirements.txt
langchain>=0.2.0
langchain-openai>=0.1.0
pandas>=2.0.0
python-dotenv>=1.0.0
pydantic>=2.0.0
tqdm>=4.65.0
tenacity>=8.0.0
pytest>=7.0.0
